requirements.txt
# 项目依赖

# 基础库
Python>=3.7

# 图形界面
Tkinter  # 标准库，无需单独安装

# 数据处理
networkx>=2.6.3  # 图结构处理
pandas>=1.3.3  # 数据处理
pyyaml>=6.0  # YAML配置文件解析

# 图像处理
Pillow>=8.3.2  # 图像处理
opencv-python>=4.5.3  # 视频处理

# 数据库
sqlite3  # 标准库，无需单独安装

# 其他
numpy>=1.21.2  # 数值计算
